Move getPostTypeLabel and usePostTypeLabel from src/helpers/event.js to src/helpers/editor.js (and their tests) since the helpers are post-type-agnostic and editor.js is the established home for WP data-store-coupled helpers.
