+++ src/helpers/event.js
@@ -87,77 +87,6 @@ export function usePostTypeSupports( support, postType = null ) {
 }
 
 /**
- * Resolve a single label from a post type's registered labels.
- *
- * JS counterpart to `Utility::post_type_label()` — wraps `getPostType()` so
- * call sites don't have to defend against unregistered post types or missing
- * label keys. Lets UI strings reflect whatever a site builder filtered the
- * labels to, and lets extenders' event-supporting post types surface their
- * own labels instead of GatherPress's defaults (#1612).
- *
- * Reads non-reactively via `select( 'core' ).getPostType()`. For React
- * components whose render output depends on the resolved label (variation
- * titles, panel headings, etc.) prefer the reactive `usePostTypeLabel`
- * hook below — without subscription the label resolves to the fallback at
- * first render and never updates.
- *
- * @since 1.0.0
- *
- * @param {string}      key      Label key to read (e.g. `singular_name`, `name`, `add_new_item`).
- * @param {string|null} postType Optional post type slug. Falls back to the editor post type.
- * @param {string}      fallback Optional fallback returned when the label can't be resolved.
- * @return {string} The resolved label, or the fallback when unresolvable.
- */
-export function getPostTypeLabel( key, postType = null, fallback = '' ) {
-	const typeToCheck =
-		postType ?? select( 'core/editor' )?.getCurrentPostType();
-
-	if ( ! typeToCheck ) {
-		return fallback;
-	}
-
-	const label = select( 'core' ).getPostType( typeToCheck )?.labels?.[ key ];
-
-	return label || fallback;
-}
-
-/**
- * Reactive variant of `getPostTypeLabel` for use in React components.
- *
- * `getPostTypeLabel` reads `getPostType()` non-reactively, so when the
- * post-type definition isn't yet cached at render time the label resolves
- * to the fallback and the component never re-renders once it loads. This
- * hook subscribes via `useSelect` so the component re-renders the moment
- * the labels become known — which is the difference between a variation
- * permanently titled "Event" and one that picks up the registered label.
- *
- * @since 1.0.0
- *
- * @param {string}      key      Label key to read.
- * @param {string|null} postType Optional post type slug. Falls back to the editor post type.
- * @param {string}      fallback Optional fallback returned when the label can't be resolved.
- * @return {string} The resolved label, or the fallback when unresolvable.
- */
-export function usePostTypeLabel( key, postType = null, fallback = '' ) {
-	return useSelect(
-		( wpSelect ) => {
-			const typeToCheck =
-				postType ?? wpSelect( 'core/editor' )?.getCurrentPostType();
-
-			if ( ! typeToCheck ) {
-				return fallback;
-			}
-
-			const label = wpSelect( 'core' ).getPostType( typeToCheck )
-				?.labels?.[ key ];
-
-			return label || fallback;
-		},
-		[ key, postType, fallback ]
-	);
-}
-
-/**
  * Checks if a post type supports event_date in the GatherPress application.
  *
  * @since 1.0.0
