+++ test/unit/js/src/helpers/event.test.js
@@ -31,8 +31,6 @@ import {
 	hasEventPastNotice,
 	isPostTypeSupporting,
 	usePostTypeSupports,
-	getPostTypeLabel,
-	usePostTypeLabel,
 	isEventPostType,
 	isRsvpPostType,
 	hasValidEventId,
@@ -224,222 +222,6 @@ describe( 'usePostTypeSupports', () => {
 		expect(
 			usePostTypeSupports( 'gatherpress-event-date', 'gatherpress_event' )
 		).toBe( true );
-	} );
-} );
-
-/**
- * Helper to create a mock getPostType function that returns labels.
- *
- * @param {string} slug The post type slug.
- * @return {Object|null} The post type object with labels.
- */
-function mockGetPostTypeWithLabels( slug ) {
-	if ( 'gatherpress_event' === slug ) {
-		return {
-			labels: {
-				name: 'Events',
-				singular_name: 'Event',
-				add_new_item: 'Add New Event',
-			},
-		};
-	}
-	if ( 'gatherpress_venue' === slug ) {
-		return {
-			labels: {
-				name: 'Venues',
-				singular_name: 'Venue',
-			},
-		};
-	}
-	return null;
-}
-
-/**
- * Coverage for getPostTypeLabel.
- */
-describe( 'getPostTypeLabel', () => {
-	it( 'returns the resolved label for the given key and post type', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'name', 'gatherpress_event' ) ).toBe( 'Events' );
-		expect( getPostTypeLabel( 'singular_name', 'gatherpress_event' ) ).toBe( 'Event' );
-		expect( getPostTypeLabel( 'add_new_item', 'gatherpress_event' ) ).toBe( 'Add New Event' );
-	} );
-
-	it( 'falls back to the editor post type when no postType is given', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => 'gatherpress_event' };
-			}
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'singular_name' ) ).toBe( 'Event' );
-	} );
-
-	it( 'returns the fallback when the post type is unknown', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'name', 'unknown_type', 'Default' ) ).toBe( 'Default' );
-	} );
-
-	it( 'returns the fallback when the requested label key is missing', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'add_new_item', 'gatherpress_venue', 'Add' ) ).toBe( 'Add' );
-	} );
-
-	it( 'returns the fallback when no post type can be resolved', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => undefined };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'name', null, 'Fallback' ) ).toBe( 'Fallback' );
-	} );
-
-	it( 'returns an empty string by default when unresolvable', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => undefined };
-			}
-			return {};
-		} );
-
-		expect( getPostTypeLabel( 'name' ) ).toBe( '' );
-	} );
-} );
-
-/**
- * Coverage for usePostTypeLabel — the reactive variant of getPostTypeLabel.
- */
-describe( 'usePostTypeLabel', () => {
-	it( 'returns the resolved label for the given key and post type', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( usePostTypeLabel( 'name', 'gatherpress_event' ) ).toBe( 'Events' );
-	} );
-
-	it( 'falls back to the editor post type when no postType is given', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => 'gatherpress_event' };
-			}
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( usePostTypeLabel( 'singular_name' ) ).toBe( 'Event' );
-	} );
-
-	it( 'returns the fallback when the post type is unknown', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		expect( usePostTypeLabel( 'name', 'unknown_type', 'Default' ) ).toBe( 'Default' );
-	} );
-
-	it( 'returns the fallback when no post type can be resolved', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => undefined };
-			}
-			return {};
-		} );
-
-		expect( usePostTypeLabel( 'name', null, 'Fallback' ) ).toBe( 'Fallback' );
-	} );
-
-	it( 'returns an empty string by default when unresolvable', () => {
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core/editor' === store ) {
-				return { getCurrentPostType: () => undefined };
-			}
-			return {};
-		} );
-
-		expect( usePostTypeLabel( 'name' ) ).toBe( '' );
-	} );
-
-	it( 'subscribes via useSelect so the label is reactive', () => {
-		// Confirms the hook delegates to useSelect — without subscription a
-		// label that resolves on a later render (post-type registry hydration)
-		// would not propagate to the component.
-		const { useSelect } = require( '@wordpress/data' );
-		useSelect.mockClear();
-		require( '@wordpress/data' ).select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return { getPostType: mockGetPostTypeWithLabels };
-			}
-			return {};
-		} );
-
-		usePostTypeLabel( 'name', 'gatherpress_event' );
-
-		expect( useSelect ).toHaveBeenCalledTimes( 1 );
-	} );
-
-	it( 're-evaluates when getPostType resolves later', () => {
-		// Simulates the actual race the hook is fixing: on first render the
-		// post-type definition isn't cached yet (returns undefined), then
-		// resolves on a subsequent invocation. The hook must reflect the new
-		// value rather than caching the fallback.
-		const { select } = require( '@wordpress/data' );
-		let resolved = false;
-		select.mockImplementation( ( store ) => {
-			if ( 'core' === store ) {
-				return {
-					getPostType: ( slug ) => {
-						if ( ! resolved ) {
-							return undefined;
-						}
-						return mockGetPostTypeWithLabels( slug );
-					},
-				};
-			}
-			return {};
-		} );
-
-		expect(
-			usePostTypeLabel( 'name', 'gatherpress_event', 'Default' )
-		).toBe( 'Default' );
-
-		resolved = true;
-
-		expect(
-			usePostTypeLabel( 'name', 'gatherpress_event', 'Default' )
-		).toBe( 'Events' );
 	} );
 } );
